backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==9.5.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.3
# Ajoutez ces lignes à votre requirements.txt
pandas>=1.5.0
openpyxl>=3.0.10  # Pour lire les fichiers Excel .xlsx
xlrd>=2.0.1       # Pour lire les fichiers Excel .xls (optionnel)